Implement payment feature
- Added payment domain entity with integer amount
- Created payment repository with count days and total used logic
- Implemented payment service with validation rules
- Added payment handler with create endpoint
- Set up payment routes under /api/v1/payments
- Added request/response DTOs for payments
- Integrated payment system into main API setup
- Fixed day routes path from /day to /days

+++ backend/internal/domain/entities/payment.go
@@ -6,7 +6,7 @@ import (
 
 type Payment struct {
 	ID          int       `gorm:"primaryKey" json:"id"`
-	Amount      float64   `gorm:"not null;default:0" json:"amount"` // Valor pago (ex: 140.00)
+	Amount      int       `gorm:"not null;default:0" json:"amount"` // Valor pago (ex: 140)
 	PaymentDate time.Time `gorm:"not null" json:"payment_date"`     // Data do pagamento
 	UserID      int       `gorm:"not null;index" json:"user_id"`    // Chave estrangeira
 	CreatedAt   time.Time `gorm:"autoCreateTime" json:"created_at"`
